+++ src/CommonITILValidation.php
@@ -55,11 +55,11 @@ abstract class CommonITILValidation extends CommonDBChild
     public const VALIDATE               = 1024;
 
 
-    // STATUS
-    public const NONE      = 1; // none
-    public const WAITING   = 2; // waiting
-    public const ACCEPTED  = 3; // accepted
-    public const REFUSED   = 4; // rejected
+    // STATUSES
+    public const NONE      = 1; // used for ticket.global_validation
+    public const WAITING   = 2;
+    public const ACCEPTED  = 3;
+    public const REFUSED   = 4;
 
     public static function getIcon()
     {
@@ -75,7 +75,6 @@ abstract class CommonITILValidation extends CommonDBChild
     {
         return [CREATE];
     }
-
 
     public static function getPurgeRights()
     {
@@ -284,6 +283,10 @@ abstract class CommonITILValidation extends CommonDBChild
 
     public function prepareInputForAdd($input)
     {
+        // validation step is mandatory : add default value is not set
+        if (!isset($input['itils_validationsteps_id']) && !isset($input['_validationsteps_id'])) {
+            $input['_validationsteps_id'] = ValidationStep::getDefault()->getID();
+        }
 
         $input["users_id"] = 0;
         // Only set requester on manual action
@@ -321,11 +324,30 @@ abstract class CommonITILValidation extends CommonDBChild
         return parent::prepareInputForAdd($input);
     }
 
+    #[Override()]
+    public function pre_addInDB()
+    {
+        if (array_key_exists('_validationsteps_id', $this->input)) {
+            $this->addITILValidationStep($this->input['_validationsteps_id']);
+        }
+    }
+
 
     public function post_addItem()
     {
         /** @var array $CFG_GLPI */
         global $CFG_GLPI;
+
+        if (isset($this->input['_validationsteps_threshold'])) {
+            $this->updateITILValidationStepThreshold((int) $this->input['_validationsteps_threshold']);
+        }
+
+        $itilobject_type = new static::$itemtype();
+        $itilobject_id = $this->fields[static::$items_id];
+
+        if (!$itilobject_type->getFromDB($itilobject_id)) {
+            throw new \RuntimeException('Item related to the validation not found ' . static::$itemtype . ' ID: ' . $itilobject_id);
+        }
 
         // Handle rich-text images
         foreach (['comment_submission', 'comment_validation'] as $content_field) {
@@ -339,60 +361,57 @@ abstract class CommonITILValidation extends CommonDBChild
         // Handle uploaded documents
         $this->input = $this->addFiles($this->input);
 
-        $item     = new static::$itemtype();
+        // --- update item (ITILObject) handling the validation
+        // always recompute global validation status on ticket
+        $input = [
+            'id' => $itilobject_id,
+            'global_validation' => static::computeValidationStatus($itilobject_type),
+            '_from_itilvalidation' => true,
+        ];
+
+        // to fix lastupdater
+        if (isset($this->input['_auto_update'])) {
+            $input['_auto_update'] = $this->input['_auto_update'];
+        }
+        // to know update by rules
+        if (isset($this->input["_rule_process"])) {
+            $input['_rule_process'] = $this->input["_rule_process"];
+        }
+        // No update ticket notif on ticket add
+        if (isset($this->input["_ticket_add"])) {
+            $input['_disablenotif'] = true;
+        }
+        $itilobject_type->update($input);
+
+        // -- send email notification
         $mailsend = false;
-        if ($item->getFromDB($this->fields[static::$items_id])) {
-            // Set global validation to waiting
-            if (((int) $item->fields['global_validation']) !== self::WAITING) {
-                $input = [
-                    'id'                    => $this->fields[static::$items_id],
-                    'global_validation'     => self::WAITING,
-                    '_from_itilvalidation'  => true,
-                ];
-
-                // to fix lastupdater
-                if (isset($this->input['_auto_update'])) {
-                    $input['_auto_update'] = $this->input['_auto_update'];
+        if (!isset($this->input['_disablenotif']) && $CFG_GLPI["use_notifications"]) {
+            $options = ['validation_id' => $this->fields["id"],
+                'validation_status' => $this->fields["status"],
+            ];
+            $mailsend = NotificationEvent::raiseEvent('validation', $itilobject_type, $options, $this);
+        }
+        if ($mailsend) {
+            if ($this->fields['itemtype_target'] === 'User') {
+                $user = new User();
+                $user->getFromDB($this->fields["items_id_target"]);
+                $email = $user->getDefaultEmail();
+                if (!empty($email)) {
+                    Session::addMessageAfterRedirect(htmlescape(sprintf(__('Approval request sent to %s'), $user->getName())));
+                } else {
+                    Session::addMessageAfterRedirect(
+                        htmlescape(sprintf(
+                            __('The selected user (%s) has no valid email address. The request has been created, without email confirmation.'),
+                            $user->getName()
+                        )),
+                        false,
+                        ERROR
+                    );
                 }
-                // to know update by rules
-                if (isset($this->input["_rule_process"])) {
-                    $input['_rule_process'] = $this->input["_rule_process"];
-                }
-                // No update ticket notif on ticket add
-                if (isset($this->input["_ticket_add"])) {
-                    $input['_disablenotif'] = true;
-                }
-                $item->update($input);
-            }
-
-            if (!isset($this->input['_disablenotif']) && $CFG_GLPI["use_notifications"]) {
-                $options = ['validation_id'     => $this->fields["id"],
-                    'validation_status' => $this->fields["status"],
-                ];
-                $mailsend = NotificationEvent::raiseEvent('validation', $item, $options, $this);
-            }
-            if ($mailsend) {
-                if ($this->fields['itemtype_target'] === 'User') {
-                    $user    = new User();
-                    $user->getFromDB($this->fields["items_id_target"]);
-                    $email   = $user->getDefaultEmail();
-                    if (!empty($email)) {
-                        Session::addMessageAfterRedirect(htmlescape(sprintf(__('Approval request sent to %s'), $user->getName())));
-                    } else {
-                        Session::addMessageAfterRedirect(
-                            htmlescape(sprintf(
-                                __('The selected user (%s) has no valid email address. The request has been created, without email confirmation.'),
-                                $user->getName()
-                            )),
-                            false,
-                            ERROR
-                        );
-                    }
-                } elseif (is_a($this->fields["itemtype_target"], CommonDBTM::class, true)) {
-                    $target = new $this->fields["itemtype_target"]();
-                    if ($target->getFromDB($this->fields["items_id_target"])) {
-                        Session::addMessageAfterRedirect(htmlescape(sprintf(__('Approval request sent to %s'), $target->getName())));
-                    }
+            } elseif (is_a($this->fields["itemtype_target"], CommonDBTM::class, true)) {
+                $target = new $this->fields["itemtype_target"]();
+                if ($target->getFromDB($this->fields["items_id_target"])) {
+                    Session::addMessageAfterRedirect(htmlescape(sprintf(__('Approval request sent to %s'), $target->getName())));
                 }
             }
         }
@@ -402,18 +421,19 @@ abstract class CommonITILValidation extends CommonDBChild
 
     public function prepareInputForUpdate($input)
     {
-        $can_answer = $this->canAnswer();
-        // Don't allow changing internal entity fields or change the item it is attached to
         $forbid_fields = ['entities_id', static::$items_id, 'is_recursive'];
         // The following fields shouldn't be changed by anyone after the approval is created
         array_push(
             $forbid_fields,
+            'itils_validationsteps_id',
             'users_id',
             'itemtype_target',
             'items_id_target',
             'submission_date'
         );
 
+        // Don't allow changing internal entity fields or change the item it is attached to
+        $can_answer = $this->canAnswer();
         if (!$can_answer) {
             array_push($forbid_fields, 'status', 'comment_validation', 'validation_date');
         }
@@ -451,13 +471,25 @@ abstract class CommonITILValidation extends CommonDBChild
         return parent::prepareInputForUpdate($input);
     }
 
+    public function post_purgeItem()
+    {
+        $this->removeUnsedITILValidationStep($this->fields['itils_validationsteps_id']);
+        $this->recomputeItilStatus();
+
+        parent::post_purgeItem();
+    }
 
     public function post_updateItem($history = true)
     {
         /** @var array $CFG_GLPI */
         global $CFG_GLPI;
 
-        $item    = new static::$itemtype();
+        if (isset($this->input['_validationsteps_threshold'])) {
+            $this->updateITILValidationStepThreshold((int) $this->input['_validationsteps_threshold']);
+        }
+
+        $this->recomputeItilStatus();
+
         $donotif = $CFG_GLPI["use_notifications"];
         if (isset($this->input['_disablenotif'])) {
             $donotif = false;
@@ -475,44 +507,47 @@ abstract class CommonITILValidation extends CommonDBChild
         // Handle uploaded documents
         $this->input = $this->addFiles($this->input);
 
-        if ($item->getFromDB($this->fields[static::$items_id])) {
-            if (
-                count($this->updates)
-                && $donotif
-            ) {
-                $options  = ['validation_id'     => $this->fields["id"],
-                    'validation_status' => $this->fields["status"],
-                ];
-                NotificationEvent::raiseEvent('validation_answer', $item, $options, $this);
-            }
-
-            //if status is updated, update global approval status
-            if (in_array("status", $this->updates)) {
-                $input = [
-                    'id'                    => $this->fields[static::$items_id],
-                    'global_validation'     => self::computeValidationStatus($item),
-                    '_from_itilvalidation'  => true,
-                ];
-                $item->update($input);
-            }
+        // -- notifications
+        if (
+            count($this->updates)
+            && $donotif
+        ) {
+            $options  = ['validation_id'     => $this->fields["id"],
+                'validation_status' => $this->fields["status"],
+            ];
+            NotificationEvent::raiseEvent('validation_answer', $this->getItem(), $options, $this);
         }
+
         parent::post_updateItem($history);
     }
 
     public function pre_deleteItem()
     {
-
         $item    = new static::$itemtype();
         if ($item->getFromDB($this->fields[static::$items_id])) {
+            // always recompute global validation status on ticket
+            // keep same behaviour for changes (since they have no validation steps)
+            if (static::$itemtype === TicketValidation::class) {
+                $input = [
+                    'id'                    => $this->fields[static::$items_id],
+                    'global_validation'     => static::computeValidationStatus($item),
+                    '_from_itilvalidation'  => true,
+                ];
+            }
+
             if (($item->fields['global_validation'] == self::WAITING)) {
                 $input = [
                     'id'                    => $this->fields[static::$items_id],
-                    'global_validation'     => self::NONE,
+                    'global_validation'     => static::computeValidationStatus($item),
                     '_from_itilvalidation'  => true,
                 ];
-                $item->update($input);
+            }
+
+            if (isset($input) && !$item->update($input)) {
+                Session::addMessageAfterRedirect(msg: 'Failed to update related ' . static::$itemtype . ' approval status  (case 3)', message_type: ERROR);
             }
         }
+
         return true;
     }
 
@@ -962,11 +997,9 @@ abstract class CommonITILValidation extends CommonDBChild
 
 
     /**
-     * Print the validation list into item
-     *
-     * @param CommonDBTM $item
-     **/
-    public function showSummary(CommonDBTM $item)
+     * Print validations summary (list of validations of the ITIL object)
+     */
+    private function showSummary(CommonITILObject $itil): void
     {
         /**
          * @var array $CFG_GLPI
@@ -984,19 +1017,17 @@ abstract class CommonITILValidation extends CommonDBChild
                 )
             )
         ) {
-            return false;
+            return;
         }
 
-        /** @var CommonITILObject $item */
-        $tID    = $item->fields['id'];
-
+        $tID    = $itil->fields['id'];
         $tmp    = [static::$items_id => $tID];
         $rand   = mt_rand();
 
-        $iterator = $DB->Request([
+        $validation_sql_results = $DB->Request([
             'FROM'   => $this->getTable(),
-            'WHERE'  => [static::$items_id => $item->getField('id')],
-            'ORDER'  => 'submission_date DESC',
+            'WHERE'  => [static::$items_id => $itil->getID()],
+            'ORDER'  => ['itils_validationsteps_id ASC', 'submission_date DESC'],
         ]);
 
         Session::initNavigateListItems(
@@ -1004,34 +1035,35 @@ abstract class CommonITILValidation extends CommonDBChild
             //TRANS : %1$s is the itemtype name, %2$s is the name of the item (used for headings of a list)
             sprintf(
                 __('%1$s = %2$s'),
-                $item::getTypeName(1),
-                $item->fields["name"]
+                $itil::getTypeName(1),
+                $itil->fields["name"]
             )
         );
-        $values = [];
-        foreach ($iterator as $row) {
-            $canedit = $this->canEdit($row["id"]);
-            Session::addToNavigateListItems($this->getType(), $row["id"]);
+        $validations = [];
+        $validationstep_id_inloop = -1;
+        foreach ($validation_sql_results as $validation) {
+            $canedit = $this->canEdit($validation["id"]);
+            Session::addToNavigateListItems($this->getType(), $validation["id"]);
             $status  = sprintf(
                 '<div class="badge fw-normal fs-4 text-wrap" style="border-color: %s;border-width: 2px;">%s</div>',
-                htmlescape(self::getStatusColor($row['status'])),
-                htmlescape(self::getStatus($row['status']))
+                htmlescape(self::getStatusColor($validation['status'])),
+                htmlescape(self::getStatus($validation['status']))
             );
 
             $comment_submission = RichText::getEnhancedHtml($this->fields['comment_submission'], ['images_gallery' => true]);
             $type_name   = null;
             $target_name = null;
-            if ($row["itemtype_target"] === User::class) {
+            if ($validation["itemtype_target"] === User::class) {
                 $type_name   = User::getTypeName();
-                $target_name = getUserName($row["items_id_target"]);
-            } elseif (is_a($row["itemtype_target"], CommonDBTM::class, true)) {
-                $target = new $row["itemtype_target"]();
+                $target_name = getUserName($validation["items_id_target"]);
+            } elseif (is_a($validation["itemtype_target"], CommonDBTM::class, true)) {
+                $target = new $validation["itemtype_target"]();
                 $type_name = $target::getTypeName();
-                if ($target->getFromDB($row["items_id_target"])) {
+                if ($target->getFromDB($validation["items_id_target"])) {
                     $target_name = $target->getName();
                 }
             }
-            $is_answered = $row['status'] !== self::WAITING && $row['users_id_validate'] > 0;
+            $is_answered = $validation['status'] !== self::WAITING && $validation['users_id_validate'] > 0;
             $comment_validation = RichText::getEnhancedHtml($this->fields['comment_validation'] ?? '', ['images_gallery' => true]);
 
             $doc_item = new Document_Item();
@@ -1053,36 +1085,14 @@ abstract class CommonITILValidation extends CommonDBChild
                 }
             }
 
-            $script = "";
+            $edit_button = "";
             if ($canedit) {
-                $edit_title = __s('Edit');
-                $item_id = (int) $item->fields['id'];
-                $row_id = (int) $row["id"];
-                $rand = htmlescape($rand);
-                $view_validation_id = htmlescape($this->fields[static::$items_id]);
-                $root_doc = htmlescape($CFG_GLPI["root_doc"]);
-                $params_json = json_encode([
-                    'type'             => static::class,
-                    'parenttype'       => static::$itemtype,
-                    static::$items_id  => $this->fields[static::$items_id],
-                    'id'               => $row["id"],
-                ]);
-
-                $script = <<<HTML
-                    <span class="ti ti-edit" style="cursor:pointer" title="{$edit_title}" 
-                          onclick="viewEditValidation{$item_id}{$row_id}{$rand}();" 
-                          id="viewvalidation{$view_validation_id}{$row_id}{$rand}">
-                    </span>
-                    <script>
-                        function viewEditValidation{$item_id}{$row_id}{$rand}() {
-                            $('#viewvalidation{$item_id}{$rand}').load('$root_doc/ajax/viewsubitem.php', $params_json);
-                        };
-                    </script>
-HTML;
+                $edit_button = $this->getModalFormHtmlElements('editapproval_modal', __('Edit Approval'), $itil->getID(), $validation["id"]) + ['anchor_title' => __('Edit') ];
             }
 
-            $values[] = [
-                'edit'                  => $script,
+            $itil_validationstep_id = $validation['itils_validationsteps_id'];
+            $validations[$itil_validationstep_id]['entries'][] = [
+                'edit'                  => $edit_button,
                 'status'                => $status,
                 'type_name'             => $type_name,
                 'target_name'           => $target_name,
@@ -1090,24 +1100,73 @@ HTML;
                 'comment_submission'    => $comment_submission,
                 'comment_validation'    => $comment_validation,
                 'document'              => $document,
-                'submission_date'       => $row["submission_date"],
-                'validation_date'       => $row["validation_date"],
-                'user'                  => getUserName($row["users_id"]),
+                'submission_date'       => $validation["submission_date"],
+                'validation_date'       => $validation["validation_date"],
+                'user'                  => getUserName($validation["users_id"]),
             ];
+
+            $itil_edit_js_identifier = 'itilvalidation_edit_' . $itil_validationstep_id;
+            $url = $itil::getValidationStepClassName()::getFormURLWithId($itil_validationstep_id);
+            $validations[$itil_validationstep_id]['edit_link_js']['js'] = Ajax::createIframeModalWindow(
+                $itil_edit_js_identifier,
+                $url,
+                [
+                    'title'           => __('Update approval step'),
+                    'reloadonclose'   => true,
+                    'display'         => false,
+                    'height'          => 120,
+                ]
+            );
+            $validations[$itil_validationstep_id]['edit_link_js']['target'] = "$itil_edit_js_identifier";
+
+            if ($itil_validationstep_id !== $validationstep_id_inloop) {
+                $itil_validationstep = $itil::getValidationStepInstance();
+                if (!$itil_validationstep->getFromDB($itil_validationstep_id)) {
+                    throw new Exception("itil Approval step not found " . $itil_validationstep_id);
+                }
+
+                $validation_step_status = $itil_validationstep->getStatus();
+                $validation_step_achievements = $itil_validationstep->getAchievements();
+
+                $vs = $itil_validationstep;
+                $vs->getFromDB($validation['itils_validationsteps_id']);
+                $validationsteps_id = $vs->fields['validationsteps_id'];
+
+                $validations[$itil_validationstep_id]['validationstep'] = [
+                    'id' => $validation['itils_validationsteps_id'],
+                    'name' => Dropdown::getDropdownName(ValidationStep::getTable(), $validationsteps_id),
+                    // structured to be later replaced by a DTO with Status::isAccepted()|isRefused()|isWaiting() & getStatus()
+                    'status' => [
+                        'waiting' => $validation_step_status === self::WAITING,
+                        'refused' => $validation_step_status === self::REFUSED,
+                        'accepted' => $validation_step_status === self::ACCEPTED,
+                    ],
+                    // structured to be later replaced by a DTO with getAchievement()
+                    'achievement' => [
+                        'waiting' => $validation_step_achievements[self::WAITING],
+                        'refused' => $validation_step_achievements[self::REFUSED],
+                        'accepted' => $validation_step_achievements[self::ACCEPTED],
+                    ],
+                    'minimal_required_validation_percent' => $itil_validationstep->fields['minimal_required_validation_percent'],
+                ];
+
+                $validationstep_id_inloop = $itil_validationstep_id;
+            }
         }
 
         TemplateRenderer::getInstance()->display('components/itilobject/validation.html.twig', [
             'canadd' => $this->can(-1, CREATE, $tmp),
-            'item' => $item,
+            'item' => $itil,
             'itemtype' => static::$itemtype,
             'tID' => $tID,
-            'donestatus' => array_merge($item->getSolvedStatusArray(), $item->getClosedStatusArray()),
+            'donestatus' => array_merge($itil->getSolvedStatusArray(), $itil->getClosedStatusArray()),
             'validation' => $this,
             'rand' => $rand,
             'items_id' => static::$items_id,
+            'sendapproval_modal' => $this->getModalFormHtmlElements('sendapproval_modal', __('Send approval request'), $itil->getID()),
         ]);
 
-        TemplateRenderer::getInstance()->display('components/datatable.html.twig', [
+        TemplateRenderer::getInstance()->display('components/sections_datatable.html.twig', [
             'is_tab' => true,
             'nopager' => true,
             'nofilter' => true,
@@ -1125,7 +1184,7 @@ HTML;
                 'document' => __('Documents'),
             ],
             'formatters' => [
-                'edit' => 'raw_html',
+                'edit' => 'html_modal',
                 'status' => 'raw_html',
                 'submission_date' => 'date',
                 'comment_submission' => 'raw_html',
@@ -1133,8 +1192,8 @@ HTML;
                 'comment_validation' => 'raw_html',
                 'document' => 'raw_html',
             ],
-            'entries' => $values,
-            'total_number' => count($values),
+            'validationsteps' => $validations,
+            'total_number' => count($validations),
             'showmassiveactions' => false,
         ]);
     }
@@ -1155,15 +1214,23 @@ HTML;
             $this->check(-1, CREATE, $options);
         }
 
+        /** @var \CommonITILObject $itil */
+        $itil = $this->getItem();
+
+
+        $ivs = $itil::getValidationStepInstance();
+        $ivs->getFromDB($this->fields['itils_validationsteps_id']);
+        $validationsteps_id = $ivs->fields['validationsteps_id'] ?? ValidationStep::getDefault()->getID();
+
         TemplateRenderer::getInstance()->display('components/itilobject/timeline/form_validation.html.twig', [
-            'item'      => $options['parent'],
-            'subitem'   => $this,
+            'item'      => $options['parent'], // ItilObject
+            'subitem'   => $this, // Validation
             'scroll'    => true,
+            '_validationsteps_id' => $validationsteps_id,
         ]);
 
         return true;
     }
-
 
     public function rawSearchOptions()
     {
@@ -1266,18 +1333,6 @@ HTML;
         $tab[] = [
             'id'                 => 'validation',
             'name'               => CommonITILValidation::getTypeName(1),
-        ];
-
-        $tab[] = [
-            'id'                 => '51',
-            'table'              => getTableForItemType(static::$itemtype),
-            'field'              => 'validation_percent',
-            'name'               => __('Minimum validation required'),
-            'datatype'           => 'number',
-            'unit'               => '%',
-            'min'                => 0,
-            'max'                => 100,
-            'step'               => 50,
         ];
 
         $tab[] = [
@@ -1776,123 +1831,22 @@ HTML;
     /**
      * Compute the validation status
      *
-     * @param $item CommonITILObject
+     * Reduced all the Validations of an item to a single status
      *
-     * @return integer
-     **/
-    public static function computeValidationStatus(CommonITILObject $item)
-    {
-
-        // Percent of validation
-        $validation_percent = $item->fields['validation_percent'];
-
-        $statuses           = [self::ACCEPTED => 0,
-            self::WAITING  => 0,
-            self::REFUSED  => 0,
-        ];
-        $validations        = getAllDataFromTable(
-            static::getTable(),
-            [
-                static::$items_id => $item->getID(),
-            ]
-        );
-
-        if ($total = count($validations)) {
-            foreach ($validations as $validation) {
-                $statuses[$validation['status']]++;
-            }
-        }
-
-        $accepted = 0;
-        $refused  = 0;
-        if ($total) {
-            $accepted = round($statuses[self::ACCEPTED] * 100 / $total);
-            $refused  = round($statuses[self::REFUSED]  * 100 / $total);
-        }
-
-        return self::computeValidation(
-            $accepted,
-            $refused,
-            $validation_percent
-        );
-    }
-
-    /**
-     * Compute the validation status from the percentage of acceptation, the
-     * percentage of refusals and the target acceptation threshold
-     *
-     * @param int $accepted             0-100 (percentage of acceptation)
-     * @param int $refused              0-100 (percentage of refusals)
-     * @param int $validation_percent   0-100 (target accepation threshold)
-     *
-     * @return int the validation status : ACCEPTED|REFUSED|WAITING
+     * @param $itil CommonITILObject
+     **@return int CommonITILValidation::VALIDATE|CommonITILValidation::REFUSED|CommonITILValidation::WAITING|CommonITILValidation::NONE
      */
-    public static function computeValidation(
-        int $accepted,
-        int $refused,
-        int $validation_percent
-    ): int {
-        if ($validation_percent > 0) {
-            if ($accepted >= $validation_percent) {
-                // We have reached the acceptation threshold
-                return self::ACCEPTED;
-            } elseif ($refused + $validation_percent > 100) {
-                // We can no longer reach the acceptation threshold
-                return self::REFUSED;
-            }
-        } else {
-            // No validation threshold set, one approval or denial is enough
-            if ($accepted > 0) {
-                return self::ACCEPTED;
-            } elseif ($refused > 0) {
-                return self::REFUSED;
-            }
-        }
-
-        return self::WAITING;
-    }
-
-
-    /**
-     * Get the validation statistics
-     *
-     * @param integer $tID tickets id
-     *
-     * @return string
-     **/
-    public static function getValidationStats($tID)
+    public static function computeValidationStatus(CommonITILObject $itil): int
     {
-
-        $tab = self::getAllStatusArray();
-
-        $nb  = countElementsInTable(static::getTable(), [static::$items_id => $tID]);
-
-        $stats = [];
-        foreach (array_keys($tab) as $status) {
-            $validations = countElementsInTable(static::getTable(), [static::$items_id => $tID,
-                'status'          => $status,
-            ]);
-            if ($validations > 0) {
-                if (!isset($stats[$status])) {
-                    $stats[$status] = 0;
-                }
-                $stats[$status] = $validations;
-            }
-        }
-
-        $list = "";
-        foreach ($stats as $stat => $val) {
-            $list .= $tab[$stat];
-            $list .= sprintf(__('%1$s (%2$d%%) '), " ", Html::formatNumber($val * 100 / $nb));
-        }
-
-        return $list;
+        $vs = $itil->getValidationStepInstance();
+        return $vs::getValidationStatusForITIL($itil);
     }
-
 
     /**
      * @param $item       CommonITILObject
      * @param $type
+     *
+     * Used in twig template
      */
     public static function alertValidation(CommonITILObject $item, $type)
     {
@@ -1992,5 +1946,153 @@ HTML;
     public static function getAllValidationStatusArray()
     {
         return [self::NONE, self::WAITING, self::REFUSED, self::ACCEPTED];
+    }
+
+    /**
+     * Elements to create a modal to edit/send an approval request
+     *
+     * @param string $dom_identifier The $itilvalidation_id parameter is happened to this identifier (unlesss it is -1)
+     * @param string $modal_title
+     * @param int $itil_id
+     * @param int $itilvalidation_id
+     * @return array{js: string, onclick: string, target: string}
+     */
+    private function getModalFormHtmlElements(string $dom_identifier, string $modal_title, int $itil_id, int $itilvalidation_id = -1): array
+    {
+        /** @var array $CFG_GLPI */
+        global $CFG_GLPI;
+
+        $dom_identifier .= $itilvalidation_id === -1 ?: (string) $itilvalidation_id;
+
+        $uri_params = [
+            'type' => static::class,
+            'parenttype' => static::$itemtype,
+            static::$items_id => $itil_id,
+            'id' => $itilvalidation_id,
+        ];
+
+        $modal['js'] = Ajax::createModalWindow(
+            $dom_identifier,
+            $CFG_GLPI['root_doc'] . "/ajax/viewsubitem.php",
+            [
+                'title'           => htmlescape($modal_title),
+                'reloadonclose'   => true,
+                'display'         => false,
+                'modal_class'     => "modal-lg",
+                'extraparams'     => $uri_params,
+            ]
+        );
+        $modal['onclick'] = 'onclick="' . $dom_identifier . '.show();"';
+        $modal['target'] = "$dom_identifier";
+
+        return $modal;
+    }
+
+    /**
+     * Associate the validation with an "itil validation step" created from an exiting "validation step"
+     *
+     * If no itils_validationsteps is defined for the itilobject, create it
+     * else, refererence it.
+     *
+     * If threshold is set, use it to create/update the itil_validationstep
+     */
+    private function addITILValidationStep(int $validationsteps_id): void
+    {
+        // find itil validations, then find if an itil_validationsteps referencing the validationstep exists
+        /** @var \CommonITILObject $itilobject_type */
+        $itilobject_type = new static::$itemtype(); // Change | Ticket
+        $itilobject_foreignkey = $itilobject_type::getForeignKeyField(); // tickets_id | changes_id
+        $v = new static();
+        $itil_id = $this->input[$itilobject_foreignkey] ?? $this->fields[$itilobject_foreignkey];
+        $itil_validations = $v->find(
+            [$itilobject_foreignkey => $itil_id,
+                ['NOT' => ['itils_validationsteps_id' => 0]],
+            ]
+        );
+        // find an itils_validationsteps related to the ticket and the validationstep
+        $itils_validationsteps_ids = array_column($itil_validations, 'itils_validationsteps_id');
+        $ivs = $itilobject_type::getValidationStepInstance();
+        $itils_validationsteps_id = $itils_validationsteps_ids[0] ?? null; // there can be only one itil_validationsteps_id related to the validationstep (or none)
+        // check if the itils_validationsteps_id is referencing the validationstep
+        $itils_validationstep_id_with_validationstep_id_exists =
+            !is_null($itils_validationsteps_id)
+            && $ivs->getFromDBByCrit(['id' => $itils_validationsteps_ids, 'validationsteps_id' => $validationsteps_id]);
+
+        if ($itils_validationstep_id_with_validationstep_id_exists) {
+            $itils_validationsteps_id = $ivs->fields['id'];
+            $this->input['itils_validationsteps_id'] = $itils_validationsteps_id;
+        } else {
+            // addValidationStep also update the current Validation
+            // load referenced ValidationStep
+            $vs = new ValidationStep();
+            if (!$vs->getFromDB($validationsteps_id)) {
+                Session::addMessageAfterRedirect('Failed to load approval step while adding approval step.');
+            };
+
+            // create ITIL_ValidationStep
+            $itil_class = static::getItilObjectItemType();
+            $itil_validationstep = $itil_class::getValidationStepInstance();
+            $itil_validationstep->add([
+                'validationsteps_id' => $validationsteps_id,
+                'minimal_required_validation_percent' => $vs->fields['minimal_required_validation_percent'],
+            ]);
+
+            $this->input['itils_validationsteps_id'] = $itil_validationstep->getID();
+        }
+
+        unset($this->input['_validationsteps_id']);
+    }
+
+    /**
+     * Delete, only if the itils_validationstep is not used anymore
+     *
+     * @param int $itils_validationsteps_id
+     * @return void
+     */
+    private function removeUnsedITILValidationStep(int $itils_validationsteps_id): void
+    {
+        /** @var \CommonITILObject $itil */
+        $itil = $this->getItem();
+        $validation = $itil::getValidationClassInstance();
+        $validations = $validation->find(['itils_validationsteps_id' => $itils_validationsteps_id]);
+        if (!empty($validations)) {
+            // itils_validation is still used, do not delete
+            return;
+        }
+
+        if (!$itil::getValidationStepInstance()->delete(['id' => $itils_validationsteps_id])) {
+            Session::addMessageAfterRedirect('Failed to delete unused approval step.');
+        };
+    }
+
+    public function recomputeItilStatus(): void
+    {
+        /** @var CommonITILObject $itil */
+        $itil = $this->getItem();
+        if (
+            !$itil->update([
+                'id' => $itil->getID(),
+                'global_validation' => self::computeValidationStatus($itil),
+                '_from_itilvalidation' => true,
+            ])
+        ) {
+            Session::addMessageAfterRedirect('Failed to update Itil global approval status.');
+        }
+    }
+
+    /**
+     * Update minimal required validation percent of ITILValidationStep
+     */
+    private function updateITILValidationStepThreshold(int $threshold): void
+    {
+        $itil_validationstep = new (static::$itemtype::getValidationStepClassName());
+        if (!$itil_validationstep->getFromDB($this->fields['itils_validationsteps_id'])) {
+            throw new \RuntimeException('Invalid ITIL validation step. ' . $this->fields['itils_validationsteps_id']);
+        };
+        if ((int) $itil_validationstep->fields['minimal_required_validation_percent'] != $threshold) {
+            if (!$itil_validationstep->update(['id' => $this->fields['itils_validationsteps_id'], 'minimal_required_validation_percent' => $threshold])) {
+                Session::addMessageAfterRedirect('Failed to update approval step threshold.');
+            };
+        }
     }
 }