+++ src/Ticket.php
@@ -3530,7 +3530,6 @@ JAVASCRIPT;
             'items_id'                  => 0,
             'locations_id'              => 0,
             'plan'                      => [],
-            'global_validation'         => CommonITILValidation::NONE,
             'time_to_resolve'           => 'NULL',
             'time_to_own'               => 'NULL',
             'slas_id_tto'               => 0,