+++ src/TicketValidation.php
@@ -39,7 +39,7 @@
 class TicketValidation extends CommonITILValidation
 {
     // From CommonDBChild
-    public static $itemtype           = 'Ticket';
+    public static $itemtype           = Ticket::class;
     public static $items_id           = 'tickets_id';
 
     public static $rightname                 = 'ticketvalidation';
@@ -106,11 +106,11 @@ class TicketValidation extends CommonITILValidation
 
         $values[self::CREATEREQUEST]
                               = ['short' => __('Create for request'),
-                                  'long'  => __('Create a validation request for a request'),
+                                  'long'  => __('Create an approval request for a request'),
                               ];
         $values[self::CREATEINCIDENT]
                               = ['short' => __('Create for incident'),
-                                  'long'  => __('Create a validation request for an incident'),
+                                  'long'  => __('Create an approval request for an incident'),
                               ];
         $values[self::VALIDATEREQUEST]
                               = __('Validate a request');